Move regression tests to their own suite
These tests are already slow and we want to be able to add dozens more.
Move them to their own suite so they are not run on every build.

Add CPU profiling so we can see where our heavy time usage is in case we
want to do some optimization.