Eliminate invalid possible states
Some states are not possible. Perhaps they duplicate a letter when they
should not. Compare each possible state with the actual tiles on the
board. If the state can't be possible with those tiles, reject it.

This eliminates our 'cannot be solved' test, as that test case generates
an invalid board state.